+++ src/main/java/io/github/simbo1905/no/framework/WriterResolver.java
@@ -1,0 +1,17 @@
+// SPDX-FileCopyrightText: 2025 Simon Massey
+// SPDX-License-Identifier: Apache-2.0
+//
+package io.github.simbo1905.no.framework;
+
+import java.util.function.Function;
+
+interface WriterResolver extends
+    Function<Class<?>, Writer> {
+  WriterResolver throwsWriterResolver = type -> {
+    throw new AssertionError("Writer throwsWriterResolver should not be reachable.");
+  };
+
+  default Writer resolveWriter(Class<?> type) {
+    return apply(type);
+  }
+}
