+++ src/test/java/io/github/simbo1905/no/framework/TypeExpr2PrimitiveArrayTests.java
@@ -1,5 +1,6 @@
 // SPDX-FileCopyrightText: 2025 Simon Massey
 // SPDX-License-Identifier: Apache-2.0
+//
 package io.github.simbo1905.no.framework;
 
 import org.junit.jupiter.api.Test;
@@ -36,7 +37,7 @@ class TypeExpr2PrimitiveArrayTests {
         new double[]{10.1, 20.2, 30.3}
     );
 
-    ComponentSerde[] serdes = Companion2.buildComponentSerdes(
+    ComponentSerde[] serdes = Companion.buildComponentSerdes(
         AllPrimitiveArrays.class,
         List.of(),
         type -> obj -> 1024, // Simplified sizer
